move product in its own header
